+++ tests/old/test_group_future_auto_qa.py
