+++ tests/old/test_group_lutz_squad.py
