+++ tests/old/test_group_optimists_of_rationality_middle_devs.py
