+++ tests/test_pandoras_box/tests/test_guide_page.py
@@ -1,4 +1,4 @@
-from tests.test_pandoras_box.pages.guide_page import GuidePage
+from tests.old.test_pandoras_box import GuidePage
 
 
 def test_tc_002_03_03_01_open_guide(driver, open_and_load_main_page):
